// AuctionPLL/src/app/generated-models/filter/complex-filter-view-model.ts
/* eslint-disable */
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { SortOrderViewModel } from './sort-order-view-model';

export class ComplexFilterViewModel {
    field: string;
    sortOrder: SortOrderViewModel;
}