// AuctionPLL/src/app/generated-models/filter/page-request-view-model.ts
/* eslint-disable */
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { CarBrandViewModel } from '../lot-models/car-brand-view-model';
import { ComplexFilterViewModel } from './complex-filter-view-model';

export class PageRequestViewModel {
    carBrand: CarBrandViewModel[];
    complexFilter: ComplexFilterViewModel[];
}